Erp_backend/requirements.txt
# Core Django & REST
django>=5.0,<6.0
djangorestframework>=3.15,<4.0

# Authentication
djangorestframework-simplejwt>=5.3.1

# CORS (frontend-backend communication)
django-cors-headers>=4.4.0

# Database driver (PostgreSQL) — remove if you stick to SQLite
psycopg2-binary>=2.9.9

# Env variables
django-environ>=0.11.2

# Filtering support for DRF (optional but useful for ERP queries)
django-filter>=24.2


#Usage
#Save this as requirements.txt in your project root.
#Install everything:

# pip install -r requirements.txt

# (Optional) Freeze exact versions for deployment:

#pip freeze > requirements.txt